requirements.txt
azure-common==1.1.23
azure-nspkg==3.0.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-nspkg==3.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cryptography==2.8
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Uploads==0.2.1
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
msrest==0.6.19
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
pycparser==2.19
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.13.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
